migrate to indy scheduled plugin
