+++ apm-agent-plugins/apm-scheduled-annotation-plugin/src/main/java/co/elastic/apm/agent/scheduled/package-info.java
@@ -23,6 +23,6 @@
  * #L%
  */
 @NonnullApi
-package co.elastic.apm.agent.spring.scheduled;
+package co.elastic.apm.agent.scheduled;
 
 import co.elastic.apm.agent.sdk.NonnullApi;
